Implemented fetching menu items from the store
