+++ internal/workbench/desktop/src/view.rs
@@ -83,63 +83,3 @@ pub enum TreeViewGroupLocation {
     PrimaryBar,
     SecondaryBar,
 }
-
-#[derive(Debug)]
-pub struct ViewsRegistry {
-    groups: HashMap<TreeViewGroupLocation, Vec<TreeViewGroup>>,
-    views: HashMap<GroupId, Vec<TreeViewDescriptor>>,
-}
-
-impl ViewsRegistry {
-    pub fn new() -> Self {
-        ViewsRegistry {
-            groups: HashMap::new(),
-            views: HashMap::new(),
-        }
-    }
-
-    pub(crate) fn append_view_group(
-        &mut self,
-        location: TreeViewGroupLocation,
-        group: TreeViewGroup,
-    ) {
-        self.groups
-            .entry(location)
-            .or_insert_with(Vec::new)
-            .push(group);
-    }
-
-    pub(crate) fn register_views(
-        &mut self,
-        id: ReadOnlyStr,
-        batch: impl IntoIterator<Item = TreeViewDescriptor>,
-    ) {
-        self.views.entry(id).or_insert_with(Vec::new).extend(batch);
-    }
-
-    pub(crate) fn get_view_descriptors_by_group_id(
-        &self,
-        id: &ReadOnlyStr,
-    ) -> Option<&Vec<TreeViewDescriptor>> {
-        self.views.get(id)
-    }
-
-    pub(crate) fn get_view_model<T: Send + Sync + Debug + 'static>(
-        &self,
-        group_id: impl Into<ReadOnlyStr>,
-        view_id: String,
-    ) -> Option<Arc<T>> {
-        self.views
-            .get(&group_id.into())?
-            .iter()
-            .find(|item| item.id == view_id)
-            .and_then(|item| Arc::downcast::<T>(Arc::clone(&item.model)).ok())
-    }
-
-    pub(crate) fn get_groups_by_location(
-        &self,
-        location: &TreeViewGroupLocation,
-    ) -> Option<&Vec<TreeViewGroup>> {
-        self.groups.get(location)
-    }
-}
