+++ view/desktop/bin/src/commands/cmd_base.rs
@@ -18,32 +18,6 @@ pub fn native_platform_info(state: State<'_, AppState>) -> NativePlatformInfo {
 }
 
 #[tauri::command]
-pub fn describe_primary_activitybar_part(
-    state: State<'_, AppState>,
-) -> Result<DescribeActivityBarPartOutput, String> {
-    let part = state
-        .workbench
-        .get_part::<PrimaryActivityBarPart>(Parts::PrimaryActivityBar.as_str())
-        .unwrap();
-
-    part.describe(state.workbench.registry())
-        .map_err(|err| format!("failed to describe primary activity bar: {err}"))
-}
-
-#[tauri::command]
-pub fn describe_primary_sidebar_part(
-    state: State<'_, AppState>,
-) -> Result<DescribeSideBarPartOutput, String> {
-    let part = state
-        .workbench
-        .get_part::<PrimarySideBarPart>(Parts::PrimarySideBar.as_str())
-        .unwrap();
-
-    part.describe(state.workbench.registry())
-        .map_err(|err| format!("failed to describe primary sidebar: {err}"))
-}
-
-#[tauri::command]
 pub fn get_view_content(state: State<'_, AppState>) -> Result<RecentsViewContent, String> {
     let model = state
         .workbench
@@ -59,13 +33,11 @@ pub fn get_view_content(state: State<'_, AppState>) -> Result<RecentsViewContent
 }
 
 #[tauri::command]
-pub fn get_menu_items(state: State<'_, AppState>) -> Option<Vec<MenuItem>> {
-    if let Some(items) = state
-        .workbench
-        .get_menu_items(&BuiltInMenuNamespaces::ViewItemContext.into())
-    {
-        Some(items.clone())
-    } else {
-        None
-    }
+pub fn get_menu_items_by_namespace(
+    state: State<'_, AppState>,
+    namespace: String,
+) -> Option<Vec<MenuItem>> {
+    state.workbench.get_menu_items_by_namespace(namespace)
 }
+
+// get_menu_items_by_many_namespaces
