+++ view/desktop/src/components/pages/Home.tsx
@@ -34,10 +34,8 @@ const SessionComponent = () => {
 
   let getAllActivities = async () => {
     try {
-      console.log((await invokeIpc("describe_primary_activitybar_part")) as object);
-      console.log((await invokeIpc("describe_primary_sidebar_part")) as object);
       // console.log((await invokeIpc("get_view_content")) as object);
-      // console.log((await invokeIpc("get_menu_items")) as object);
+      console.log((await invokeIpc("get_menu_items_by_namespace", { namespace: "headItem" })) as object);
     } catch (err) {
       console.error("Failed to get workbench state:", err);
     }
